Simplify test data loader to just be an enum
