+++ src/test/java/games/strategy/engine/data/AllianceTrackerTest.java
@@ -3,13 +3,7 @@ package games.strategy.engine.data;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.InputStream;
-import java.net.URL;
-import java.util.concurrent.atomic.AtomicReference;
-
-import games.strategy.triplea.xml.LoadGameUtil;
+import games.strategy.triplea.xml.TestMapGameData;
 import org.junit.Before;
 import org.junit.Test;
 
@@ -20,7 +14,7 @@ public class AllianceTrackerTest {
 
   @Before
   public void setUp() throws Exception {
-    m_data = LoadGameUtil.TestMapXml.TEST.getGameData();
+    m_data = TestMapGameData.TEST.getGameData();
   }
 
   @Test
