+++ src/test/java/games/strategy/triplea/oddsCalculator/ta/OddsCalculatorTest.java
@@ -21,14 +21,14 @@ import games.strategy.engine.data.Territory;
 import games.strategy.engine.data.Unit;
 import games.strategy.triplea.delegate.GameDataTestUtil;
 import games.strategy.triplea.delegate.TerritoryEffectHelper;
-import games.strategy.triplea.xml.LoadGameUtil;
+import games.strategy.triplea.xml.TestMapGameData;
 
 public class OddsCalculatorTest {
   private GameData m_data;
 
   @Before
   public void setUp() throws Exception {
-    m_data = LoadGameUtil.TestMapXml.REVISED.getGameData();
+    m_data = TestMapGameData.REVISED.getGameData();
   }
 
   @Test
@@ -87,7 +87,7 @@ public class OddsCalculatorTest {
   @Test
   public void testDefendingTransports() throws Exception {
     // use v3 rule set
-    m_data = LoadGameUtil.TestMapXml.WW2V3_1942.getGameData();
+    m_data = TestMapGameData.WW2V3_1942.getGameData();
     final Territory sz1 = territory("1 Sea Zone", m_data);
     final List<Unit> attacking = submarine(m_data).create(2, americans(m_data));
     final List<Unit> defending = transport(m_data).create(2, germans(m_data));
